move class - NotificationPreference to member package
