+++ src/main/java/co/fineants/api/domain/common/notification/Notifiable.java
@@ -2,7 +2,7 @@ package co.fineants.api.domain.common.notification;
 
 import co.fineants.api.domain.notification.domain.dto.response.NotifyMessage;
 import co.fineants.api.domain.notification.repository.NotificationSentRepository;
-import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
+import co.fineants.api.domain.member.domain.entity.NotificationPreference;
 
 public interface Notifiable {
 	Long fetchMemberId();
