+++ src/main/java/co/fineants/api/domain/member/domain/dto/response/MemberNotificationPreferenceResponse.java
@@ -1,6 +1,6 @@
 package co.fineants.api.domain.member.domain.dto.response;
 
-import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
+import co.fineants.api.domain.member.domain.entity.NotificationPreference;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
