+++ src/main/java/co/fineants/api/domain/notification/config/NotificationConfig.java
@@ -12,7 +12,7 @@ import co.fineants.api.domain.notification.domain.entity.policy.MaxLossNotificat
 import co.fineants.api.domain.notification.domain.entity.policy.TargetGainNotificationPolicy;
 import co.fineants.api.domain.notification.domain.entity.policy.TargetPriceNotificationPolicy;
 import co.fineants.api.domain.notification.repository.NotificationSentRepository;
-import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
+import co.fineants.api.domain.member.domain.entity.NotificationPreference;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
