+++ src/main/java/co/fineants/api/domain/notification/domain/entity/policy/AbstractNotificationPolicy.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.function.Predicate;
 
 import co.fineants.api.domain.common.notification.Notifiable;
-import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
+import co.fineants.api.domain.member.domain.entity.NotificationPreference;
 
 public abstract class AbstractNotificationPolicy implements NotificationPolicy<Notifiable> {
