+++ src/test/java/co/fineants/api/domain/member/service/MemberNotificationPreferenceServiceTest.java
@@ -16,7 +16,7 @@ import co.fineants.api.domain.member.domain.dto.request.MemberNotificationPrefer
 import co.fineants.api.domain.member.domain.dto.response.MemberNotificationPreferenceResponse;
 import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.domain.member.repository.MemberRepository;
-import co.fineants.api.domain.notificationpreference.domain.entity.NotificationPreference;
+import co.fineants.api.domain.member.domain.entity.NotificationPreference;
 import co.fineants.api.domain.notificationpreference.repository.NotificationPreferenceRepository;
 import co.fineants.api.global.errors.exception.business.ForbiddenException;
